refactor: convert static method to utility function
